bug fix + healthbar init

+++ Source/Curse_of_the_Abyss/Sprites/MovableSprite.cs
@@ -7,7 +7,7 @@ namespace Curse_of_the_Abyss
     public class MovableSprite : Sprite
     {
         protected double xVelocity;
-        protected float yVelocity;
+        protected double yVelocity;
         protected float xAcceleration;
         protected float yAcceleration;
         public MovableSprite()
